Add files via upload
Above includes the html and ipynb of my first draft. The total.fasta includes all of the sequences needed for alignment of the virus strains. The pdb files are all of the files necessary for the 3D protein structures of the virus strains. 
Scientific Question: Why does the hemagglutinin (HA) in the influenza A H1N1, and H3N2 and the influenza B Yamagata and Victoria viruses affect the protection provided from the influenza vaccine?
Hypothesis: Since the influenza vaccine incorporates all four of the virus strains, it is known that the vaccine provides the least amount of protection against influenza A H3N2 than the rest. Based on this previous research, if humans are exposed to influenza a H1N1, influenza a H3N2, influenza B Yamagata, and influenza B Victoria, then the influenza vaccine will provide the least protection against H3N2 because the hemagglutinin (HA) of the H3N2 virus will vary the most from H1N1, Yamagata and Victoria strains.